State rules, use delegation to meet the first rule
In this refactoring, I'm going to follow the rules of East Oriented
Code (details here
http://saturnflyer.com/blog/jim/2014/12/23/enforcing-encapsulation-with-east-oriented-code/
and here
http://www.confreaks.com/videos/4825-RubyConf2014-eastward-ho-a-clear-path-through-ruby-with-oo)

Just for the sake of all of our memories, I've put the rules at the top
of the code.

The first rule is to return self. Now, the `convert` method can't
return self because it's a class method. But I'm going to treat
`convert` as a declarative builder
(http://programming.ianwhitney.com/blog/2014/04/13/4-simple-rules-and-declarative-builders/)
meaning that I expect it to return an instance of Raindrops.

To do that, I need to be able to instantiate instances of Raindrops and
have them be string-like enough that the tests will pass. I'm going to
do that with simple delegator. I do not need to explicitly override the
initialize method, but I'm doing so to illustrate what's going on and to
show that I am explicitly returning self