Wrap Prime with Factors
Our use of Prime didn't follow the East Orientation rules, since it was
a query on another object.

For the purposes of this refactoring, I'm going to wrap that logic in a
new class that does follow the rules, Factors. Factors still works with
Prime, but this hides the implementation from Raindrops