Extract sound construction to RainSounds
`convert` queried the return value of Factors extensively, which
violates rule #2. The first step to fixing this is to extract the logic
into a new class, RainSounds (which, yes, terrible name). This
class returns a constructed string, or an empty string. Querying of
numbers is still happening inside of RainSounds.new, which we'll tackle
next.