Express intent by Extracting Class
Rule 1 passes, so we can ignore it. Rule 2, though. I don't think this
code expresses its intent. In the first line we're converting a number
to an array, but the reason for doing so is unclear. I think this is a
problem of naming, mostly.

I'm trying to make this intent more clear by creating a class and giving
it an intent-revealing name: DropCounts. To me this more clearly
expresses what this thing represents.

Part of the problem with this exercise is that it is hard to fit into a
mental model. All it says is that if you get a number that has a prime
factor of 3, then say Pling. But what does that 3 represent? Why does 3
mean Pling? It's unclear. I'm trying to add some clarity by saying that
3 is a number of drops that land.

The refactoring I've done here is Extract Class (page 149 in
Refactoring). This new class returns an enumerable that behaves exactly
the same as the old implementation, so no other code changes are
necessary.