Express intent with extract of Sound class
Similar to the extraction of DropCounts, this extraction of a class
helps clarify our intent.

We had logic that created a sound. It was all related, so it was easy to
extract into its own Sound class. As before, we're using the Extract
Class refactoring.

The `made_by` constructor follows another suggestion in Corey Haines'
book, which I call Declarative Builders. Instead of just using `new`,
use a constrcutor that more clearly states what you're constructing.

I didn't use this approach in DropCounts because I couldn't think of a
good constructor name.