Encapsulate Field, maybe?
This might be personal preference, or it might be an actual refactoring
from the book. The line is blurry.

What I'm trying to do here is separate the logic that sets the default
value of the sound as an empty string. Mostly I'm doing that for future
refactorings. I don't think there's any particular smell that is making
me do so now.

This use of a private method that works as both a getter & setter is a
common approach that I take for logic like this.