Remove Duplication and Express Intent with Surfaces
More than one refactoring in this commit, since I didn't see a way to do
this in smaller steps.

I'm trying to tackle two problems here. First, Sound was looking a lot
like a Factory. I give it a number, it returns a string. But I didn't
see any reason for it to know the details of whot goes into that string.
That seemed to be someone else's responsibility

So, stealing an idea from Jim Gay (in a discussion we had on this gist:
https://gist.github.com/IanWhitney/6d8d777659896ff9e20d) I'm adding a
"Surface" to the mix.

At the highest level, `Raindrops.convert`, I'm setting a default surface
of Bucket. This is basically the "Introduce Parameter Object"
refactoring, though we skipped a step. The intermediary step might have
looked like

```
def self.convert(number, surface = {3 => "Pling", 5 => "Plang", 7 =>"Plong"}
  ...
end
```
This is because "Introduce Parameter Object" assumes that you're passing
in the parameters before you extract them out to an object.

I've given surface the ability to make a sound when struck. This removes
that knowledge from Sound. This seems like the right organization and is
an example of Replace Conditional with Polymorphism, though so far there
is only one type of Surface