Rename class, move responsibilities
I've never liked the conditional in Raindrops.convert that returns
either the sound or the number as a string. First, it's a sneaky
duplication of knowledge. Sound is a string, so I have to also convert
the number to string before returning it. That seems like something that
will surprise me later.

Second, it doesn't seem like something that raindrops should know. As
I've gone through this refactoring, the logic for building the sound has
moved more and more into Sound. So maybe Sound should handle this
conditional?

And when I move this logic into sound, I can wrap it all up in the to_s
method. Actually, having it in the to_s method is probably terrible,
because if the class is used in unexpected ways, the behavior will be
weird.

Along with this change, there's a name change. Sound becomes SoundEffect
because I thought it was more descriptive. Like I said earlier, this
code is kinda goofy when you try to map it on to real-world things.

And I'm injecting RaindropCounts as a dependency, instead of having
Raindrops have hard-coded knowledge about it.